created new components for tour sections, broke functions and types into seperate files
